api/requirements.txt
fastapi~=0.86.0
nltk~=3.7
gensim~=4.2.0
PyYAML~=6.0
uvicorn~=0.19.0